+++ app/src/main/java/com/tasomaniac/openwith/settings/advanced/features/ToggleFeatureActivity.kt
@@ -1,4 +1,4 @@
-package com.tasomaniac.openwith.settings.advanced
+package com.tasomaniac.openwith.settings.advanced.features
 
 import android.app.Activity
 import android.content.Intent
