+++ app/src/main/res/xml/pref_features.xml
@@ -4,16 +4,16 @@
   <Preference
     android:key="@string/pref_key_feature_add_to_homescreen"
     android:title="@string/pref_title_feature_add_to_homescreen"
-    android:fragment="com.tasomaniac.openwith.settings.advanced.ToggleFeatureActivity"/>
+    android:fragment="com.tasomaniac.openwith.settings.advanced.features.ToggleFeatureActivity"/>
 
   <Preference
     android:key="@string/pref_key_feature_text_selection"
     android:title="@string/pref_title_feature_text_selection"
-    android:fragment="com.tasomaniac.openwith.settings.advanced.ToggleFeatureActivity"/>
+    android:fragment="com.tasomaniac.openwith.settings.advanced.features.ToggleFeatureActivity"/>
 
   <Preference
     android:key="@string/pref_key_feature_direct_share"
     android:title="@string/pref_title_feature_direct_share"
-    android:fragment="com.tasomaniac.openwith.settings.advanced.ToggleFeatureActivity"/>
+    android:fragment="com.tasomaniac.openwith.settings.advanced.features.ToggleFeatureActivity"/>
 
 </PreferenceScreen>
